requirements.txt
langchain>=0.1.0
langgraph>=0.0.24
langchain-google-genai>=1.0.1
langchain-core>=0.1.0
pandas>=2.2.0
python-dotenv>=1.0.1
rich>=13.0.0
tenacity>=8.2.0